chore: semplificato il tipo per gli importi

+++ 04_driven_pull/src/main/java/io/doubleloop/drivenpull/BalanceResult.java
@@ -1,14 +1,13 @@
 package io.doubleloop.drivenpull;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class BalanceResult {
   public final String userId;
   public final LocalDate date;
-  public final BigDecimal balance;
+  public final Double balance;
 
-  public BalanceResult(String userId, LocalDate date, BigDecimal balance) {
+  public BalanceResult(String userId, LocalDate date, Double balance) {
     this.userId = userId;
     this.date = date;
     this.balance = balance;
